requirements.txt
Flask==2.3.3
ultralytics==8.0.196
opencv-python-headless==4.8.1.78
Pillow==10.0.1
numpy==1.24.3
gunicorn==21.2.0
torch==2.1.0
torchvision==0.16.0